dev: refactor per page, use context
